Context is now based on an interface, instead of a class.
By using an interface, curveball plugins are able to extend the
interface.
