+++ src/index.ts
@@ -1,4 +1,5 @@
 import { default as Application, invokeMiddlewares, Middleware, middlewareCall } from './application';
+import BaseContext from './base-context';
 import Context from './context';
 import Headers from './headers';
 import MemoryRequest from './memory-request';
@@ -9,6 +10,7 @@ import Response from './response';
 export default Application;
 export {
   Application,
+  BaseContext,
   Context,
   Headers,
   invokeMiddlewares,
